Revert "Cat 601b"
