+++ src/App.css
@@ -375,10 +375,6 @@ footer a {
   width: max-content;
 }
 
-.cat-vh-40 {
-  height: 40vh;
-}
-
 .cat-vh-60 {
   height: 60vh;
 }
@@ -487,11 +483,4 @@ a.plain:hover {
 
 .flex {
   display: flex !important;
-}
-
-.circle-70-grey {
-  width: 70px;
-  text-align: center;
-  border: 3px solid #d3d3d3;
-  border-radius: 100%;
 }
