+++ src/App.tsx
@@ -35,9 +35,6 @@ import PrincipleDetails from "./pages/principles/PrincipleDetails";
 import MotivationActorCriteria from "./pages/motivations/MotivationActorCriteria";
 import AssessmentView from "./pages/assessments/AssessmentView";
 import { MotivationAssessmentEditor } from "./pages/assessments/MotivationAssessmentEditor";
-import MotivationCriteriaPrinciples from "./pages/motivations/MotivationCriteriaPrinciples";
-import Criteria from "./pages/criteria/Criteria";
-import CriterionDetails from "./pages/criteria/CriterionDetails";
 
 const queryClient = new QueryClient();
 
@@ -221,12 +218,6 @@ function App() {
                     <Route index element={<MotivationDetails />} />
                   </Route>
                   <Route
-                    path="/motivations/:mtvId/manage-criteria-principles"
-                    element={<ProtectedRoute />}
-                  >
-                    <Route index element={<MotivationCriteriaPrinciples />} />
-                  </Route>
-                  <Route
                     path="/motivations/:mtvId/actors/:actId"
                     element={<ProtectedRoute />}
                   >
@@ -243,12 +234,6 @@ function App() {
                   </Route>
                   <Route path="/principles/:id" element={<ProtectedRoute />}>
                     <Route index element={<PrincipleDetails />} />
-                  </Route>
-                  <Route path="/criteria" element={<ProtectedRoute />}>
-                    <Route index element={<Criteria />} />
-                  </Route>
-                  <Route path="/criteria/:id" element={<ProtectedRoute />}>
-                    <Route index element={<CriterionDetails />} />
                   </Route>
                   <Route path="/login" element={<ProtectedRoute />}>
                     <Route index element={<Profile />} />
