+++ src/api/index.ts
@@ -6,6 +6,3 @@ export * from "./services/users";
 export * from "./services/organisations";
 export * from "./services/templates";
 export * from "./services/validations";
-export * from "./services/motivations";
-export * from "./services/principles";
-export * from "./services/criteria";
