+++ src/config.json
@@ -8,11 +8,6 @@
   },
   "relation_ids": {
     "motivation_principle": "isSupportedBy",
-    "motivation_actor": "dcterms:isRequiredBy",
-    "motivation_principle_criterion": "maintainedBy"
-  },
-  "default_values": {
-    "criterion_type": "Criterion",
-    "criterion_imperative": "MUST"
+    "motivation_actor": "dcterms:isRequiredBy"
   }
 }
