+++ src/pages/motivations/MotivationDetails.tsx
@@ -172,15 +172,6 @@ export default function MotivationDetails() {
             >
               Update Details
             </Link>
-            {motivation !== undefined && (
-              <Link
-                id="manage-motivation-criteria-principles"
-                to={`/motivations/${motivation.id}/manage-criteria-principles`}
-                className="btn btn-dark mt-4"
-              >
-                Manage Criteria & Principles
-              </Link>
-            )}
           </div>
         </Col>
         <Col className="col-md-auto col-lg-9 border-right">
@@ -444,7 +435,7 @@ export default function MotivationDetails() {
         <Button
           variant="secondary"
           onClick={() => {
-            navigate(`/motivations`);
+            navigate(-1);
           }}
         >
           Back
